+++ R/impl-sql.R
@@ -656,7 +656,12 @@ read_report_data_csv <- function (csv, env, dict, time,
                         day_type = day_type)
                 }
 
+                # add time columns for subsetting
+                cols <- c("environment_period_index", "month", "day", "hour", "minute")
+                add_joined_cols(time, data, "time_index", cols)
                 data <- subset_by_interval(data, int_var)
+                set(data, NULL, cols, NULL)
+
                 set(data, NULL, "time_index", time_sub$time_index)
             }
         }
@@ -685,6 +690,7 @@ read_report_data_csv <- function (csv, env, dict, time,
 
         melt_data <- function (data, int_per, dict_per, time_per) {
             if (!nrow(time_per)) return(data.table(time_index = integer(), report_data_dictionary_index = integer(), value = double()))
+            if (!nrow(dict_per)) return(data.table(time_index = integer(), report_data_dictionary_index = integer(), value = double()))
 
             cols <- setdiff(names(data), setdiff(dict$Variable, dict_per$Variable))
             data <- subset_by_interval(fast_subset(data, cols), int_per, timestep)